+++ app/src/main/resources/db-migration/00026__schema.sql
@@ -1,3 +1,6 @@
-alter table player add column localLlmAddress text default 'http://localhost:11434';
-alter table game_session add column localLlmCommandModel text default 'qwen2.5:14b';
-alter table game_session add column localLlmQueryModel text default 'qwen2.5:14b';
+alter table player
+    add column localLlmAddress text default 'http://localhost:1234/v1/chat/completions';
+alter table game_session
+    add column localLlmCommandModel text default 'matrixportalx/tulu-3.1-8b-supernova';
+alter table game_session
+    add column localLlmQueryModel text default 'matrixportalx/tulu-3.1-8b-supernova';