+++ app/src/main/resources/db-migration/00030__schema.sql
@@ -1,3 +1,6 @@
-alter table game_session add column useLocalCommandLlm boolean default false;
-alter table game_session add column useLocalQueryLlm boolean default false;
-alter table game_session add column useLocalTTS boolean default false;
+alter table game_session
+    add column useLocalCommandLlm boolean default true;
+alter table game_session
+    add column useLocalQueryLlm boolean default true;
+alter table game_session
+    add column useLocalTTS boolean default true;