+++ app/src/main/resources/db-migration/00036__schema.sql
@@ -5,5 +5,5 @@ update game_session
 set speechSpeed = 0.30
 where id = 1;
 update player
-set localLlmAddress = 'http://localhost:11434/api/chat'
+set localLlmAddress = 'http://localhost:1234/v1/chat/completions'
 where id = 1;